+++ lib/dal/src/deprecated_action/prototype.rs
@@ -263,7 +263,10 @@ impl ActionPrototype {
         Ok(prototypes)
     }
 
-    pub async fn get_by_id(ctx: &DalContext, id: ActionPrototypeId) -> ActionPrototypeResult<Self> {
+    pub async fn get_by_id_or_error(
+        ctx: &DalContext,
+        id: ActionPrototypeId,
+    ) -> ActionPrototypeResult<Self> {
         let workspace_snapshot = ctx.workspace_snapshot()?;
         let ulid: ::si_events::ulid::Ulid = id.into();
         let node_index = workspace_snapshot.get_node_index_by_id(ulid).await?;
