+++ lib/dal/src/func.rs
@@ -12,7 +12,6 @@ use crate::change_set::ChangeSetError;
 use crate::func::argument::FuncArgumentId;
 use crate::func::intrinsics::IntrinsicFunc;
 use crate::layer_db_types::{FuncContent, FuncContentV1};
-use crate::schema::variant::SchemaVariantResult;
 use crate::workspace_snapshot::edge_weight::{
     EdgeWeightError, EdgeWeightKind, EdgeWeightKindDiscriminants,
 };
@@ -20,10 +19,7 @@ use crate::workspace_snapshot::graph::WorkspaceSnapshotGraphError;
 use crate::workspace_snapshot::node_weight::category_node_weight::CategoryNodeKind;
 use crate::workspace_snapshot::node_weight::{FuncNodeWeight, NodeWeight, NodeWeightError};
 use crate::workspace_snapshot::WorkspaceSnapshotError;
-use crate::{
-    implement_add_edge_to, pk, DalContext, HelperError, SchemaVariantId, Timestamp,
-    TransactionsError,
-};
+use crate::{implement_add_edge_to, pk, DalContext, HelperError, Timestamp, TransactionsError};
 
 use self::backend::{FuncBackendKind, FuncBackendResponseType};
 
@@ -526,33 +522,6 @@ impl Func {
         }
 
         Ok(funcs)
-    }
-
-    pub async fn list_schema_variants_for_auth_func(
-        ctx: &DalContext,
-        func_id: FuncId,
-    ) -> SchemaVariantResult<Vec<SchemaVariantId>> {
-        let workspace_snapshot = ctx.workspace_snapshot()?;
-
-        let mut schema_variant_ids = vec![];
-
-        for node_id in workspace_snapshot
-            .incoming_sources_for_edge_weight_kind(
-                func_id,
-                EdgeWeightKindDiscriminants::AuthenticationPrototype,
-            )
-            .await?
-        {
-            schema_variant_ids.push(
-                workspace_snapshot
-                    .get_node_weight(node_id)
-                    .await?
-                    .id()
-                    .into(),
-            )
-        }
-
-        Ok(schema_variant_ids)
     }
 
     /// Checks if the [`Func`] is "revertible".
